Morley.Client.AbstractClient

Documentation

class (Monad m, MonadThrow m) => HasTezosNodeRpc m where

Type class which provides basic low-level tezos-node functionality

Methods

getHeadBlock :: m Text

Get hash of the current head block, this head hash is used in other RPC calls.

getCounter :: Address -> m TezosInt64

Get address counter, which is required for both transaction sending and contract origination.

getBlockConstants :: Text -> m BlockConstants

Get block constants that required by other RPC calls.

runOperation :: RunOperation -> m RunOperationResult

Perform operation run, this operation doesn't require proper signing. As a result it returns burned gas and storage diff (also list of originated contracts but their addresses are incorrect due to the fact that operation could be not signed properly) or indicates about operation failure.

preApplyOperations :: [PreApplyOperation] -> m [RunOperationResult]

Preapply list of operations, each operation has to be signed with sender secret key. As a result it returns list of results each of which has information about burned gas, storage diff size and originated contracts.

forgeOperation :: ForgeOperation -> m HexJSONByteString

Forge operation in order to receive its hexadecimal representation.

injectOperation :: Text -> m Text

Inject operation, note that this operation has to be signed before injection. As a result it returns operation hash.

getContractScript :: Address -> m OriginationScript

Get code and storage of the desired contract. Note that both code and storage are presented in low-level Micheline representation.

waitForOperationInclusion :: Text -> m Text

Wait until operation known by some hash is included into the chain.

getContractBigMap :: Address -> GetBigMap -> m GetBigMapResult

Get big map value by contract address.