Morley.Client.Common

Description

Module with functions that used in both transaction sending and contract origination.

Documentation

data OperationConstants

Datatype that contains various values required for chain operations.

Constructors

OperationConstants

Fields

ocLastBlockHash :: Text

Block in which operations is going to be injected

ocBlockConstants :: BlockConstants

Information about block: chain_id and protocol

ocCounter :: TezosInt64

Sender counter

addOperationPrefix :: ByteString -> ByteString

getAppliedResults :: HasTezosNodeRpc m => Either RunOperation PreApplyOperation -> m [AppliedResult]

Perform runOperation or preApplyOperations and combine the results.

preProcessOperation :: HasTezosNodeRpc m => Address -> m OperationConstants

Preprocess chain operation in order to get required constants.

stubSignature :: Signature