Morley.Client.Getters

Description

Some read-only RPC functions.

Documentation

data ValueDecodeFailure

Failed to decode received value to the given type.

Constructors

ValueDecodeFailure Text T

data ValueNotFound

Constructors

ValueNotFound

readContractBigMapValue :: forall k v m. (PackedValScope k, UnpackedValScope v, HasTezosNodeRpc m) => Address -> Value k -> m (Value v)

Read big_map value of given contract by key.

If the contract contains several big_maps with given key type, only one of them will be considered.