Morley.Client.Transaction

Documentation

runTransactions :: forall t m. (HasTezosNodeRpc m, ParameterScope t) => SenderInfo -> Address -> [EntrypointParam t] -> m (Text, Text)

Perform sequence of transactions to the contract, each transactions should be defined via EntrypointParam t. Returns operation hash and block in which this operation appears.

data LEntrypointParam t

Constructors

LEntrypointParam

Fields

lepMbName :: Maybe Text

lepParam :: t

lRunTransactions :: forall m t. (HasTezosNodeRpc m, NiceParameter t) => SenderInfo -> Address -> [LEntrypointParam t] -> m (Text, Text)

Lorentz version of runTransactions